anheumai-dparys/select/select.go
package select_pkg

import (
	"fmt"
	"net/http"
	"time"
)

var tenSecondsTimeOut = 10 * time.Second

func oldRacer(a, b string) string {

	aDuration := measureResponseTime(a)
	bDuration := measureResponseTime(b)

	if aDuration < bDuration {
		return a
	}

	return b
}

func Racer(a, b string) (string, error) {
	return ConfigurableRacer(a, b, tenSecondsTimeOut)
}

func ConfigurableRacer(a, b string, timeout time.Duration) (string, error) {
	select {
	case <-ping(a):
		return a, nil
	case <-ping(b):
		return b, nil
	case <-time.After(timeout):
		return "", fmt.Errorf("time out happened")
	}
}

func ping(url string) chan struct{} {
	ch := make(chan struct{})
	go func() {
		http.Get(url)
		close(ch)
	}()
	return ch

}

func measureResponseTime(url string) time.Duration {
	start := time.Now()
	http.Get(url)
	return time.Since(start)
}
